// JohnBryce-Lessons/2026.01.06/02 - MONGO + NODEJS/frontend/src/app/core/models/user.model.ts
export interface User {
  id: string;
  name: string;
  email: string;
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export interface RegisterData {
  name: string;
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}